Implement basic API
This codebase should now implement a very basic version of the final product.
At the moment it just throws back the JSON from the graph API, and some tweaks
may be needed to that, in order to ensure the list is completely scanned, and
that exactly the desired fields are included, but otherwise, that part is
almost done.

This also included some work on the page, some of which may end up being able
to be undone. I don't know how much will need to be shown on the page, it
should be able to be very simple.